Changed Map to DisplayMap, as to not confuse with Map component from 'react-leaflet'
